refactor(console): sie details page
