+++ packages/console/src/pages/SignInExperience/tabs/Others/AuthenticationForm.tsx
@@ -1,6 +1,7 @@
 import { useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 
+import Card from '@/components/Card';
 import FormField from '@/components/FormField';
 import Switch from '@/components/Switch';
 
@@ -12,7 +13,7 @@ const AuthenticationForm = () => {
   const { register } = useFormContext<SignInExperienceForm>();
 
   return (
-    <>
+    <Card>
       <div className={styles.title}>{t('sign_in_exp.others.advanced_options.title')}</div>
       <FormField title="sign_in_exp.others.advanced_options.enable_user_registration">
         <Switch
@@ -20,7 +21,7 @@ const AuthenticationForm = () => {
           label={t('sign_in_exp.others.advanced_options.enable_user_registration_description')}
         />
       </FormField>
-    </>
+    </Card>
   );
 };
