+++ packages/console/src/pages/SignInExperience/tabs/SignUpAndSignIn/SignInForm.tsx
@@ -1,15 +1,16 @@
 import { useTranslation } from 'react-i18next';
 
+import Card from '@/components/Card';
 import FormField from '@/components/FormField';
 
+import * as styles from '../index.module.scss';
 import SignInMethodEditBox from './components/SignInMethodEditBox';
-import * as styles from './index.module.scss';
 
 const SignInForm = () => {
   const { t } = useTranslation(undefined, { keyPrefix: 'admin_console' });
 
   return (
-    <>
+    <Card>
       <div className={styles.title}>{t('sign_in_exp.sign_up_and_sign_in.sign_in.title')}</div>
       <FormField title="sign_in_exp.sign_up_and_sign_in.sign_in.sign_in_identifier_and_auth">
         <div className={styles.formFieldDescription}>
@@ -17,7 +18,7 @@ const SignInForm = () => {
         </div>
         <SignInMethodEditBox />
       </FormField>
-    </>
+    </Card>
   );
 };
