--- packages/console/src/pages/SignInExperience/tabs/SignUpAndSignIn/index.module.scss
@@ -1,28 +1,0 @@
-@use '@/scss/underscore' as _;
-
-.title {
-  @include _.subhead-cap;
-  color: var(--color-neutral-variant-60);
-  margin-top: _.unit(12);
-
-  &:first-child {
-    margin-top: _.unit(6);
-  }
-}
-
-.formFieldDescription {
-  font: var(--font-body-medium);
-  color: var(--color-text-secondary);
-  margin: _.unit(1) 0 _.unit(2);
-}
-
-.socialOnlyDescription {
-  margin-left: _.unit(1);
-  color: var(--color-text-secondary);
-}
-
-.selections {
-  > :not(:first-child) {
-    margin-top: _.unit(3);
-  }
-}
